[JENKINS-69187] Split and Deprecate SupportAction methods
