+++ src/main/resources/com/cloudbees/jenkins/support/SupportAction/action.jelly
@@ -24,6 +24,6 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:l="/lib/layout">
   <l:task icon="${h.getIconFilePath(action)}" title="${action.displayName}"
-          href="${h.getActionUrl(it.url,action)}/generateSupportBundle"
+          href="${h.getActionUrl(it.url,action)}/${action.defaultChildAction.urlName}"
           permission="${app.ADMINISTER}"/>
 </j:jelly>
