--- src/main/resources/com/cloudbees/jenkins/support/SupportAction/generatedSupportContent.properties
@@ -1,2 +1,0 @@
-GeneratedSupportContent=Generated Support Content
-ExistingSupportBundles=Existing Support bundles
