+++ src/main/resources/com/cloudbees/jenkins/support/SupportAction/index.jelly
@@ -1,7 +1,7 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:l="/lib/layout">
   <l:layout title="${it.actionTitleText}" norefresh="true" permission="${app.ADMINISTER}">
-    <st:include page="sidepanel.jelly"/>
+    <st:include page="sidepanel"/>
     <l:main-panel/>
   </l:layout>
 </j:jelly>
