+++ src/main/resources/com/cloudbees/jenkins/support/actions/Messages.properties
@@ -22,6 +22,12 @@
 # THE SOFTWARE.
 #
 
+SupportGenerateBundleAction_DisplayName=Generate a Support Bundle
+SupportGenerateBundleAction_DefaultActionTitle=Generate a Support Bundle
+
+SupportGeneratedContentAction_DisplayName=Manage Generated Content
+SupportGeneratedContentAction_DefaultActionTitle=Manage Generated Content
+
 SupportRunAction_DisplayName={0} Support
 SupportRunAction_DefaultActionTitle=Support
 SupportRunAction_CreateBundle=Generate bundle
