+++ src/main/resources/com/cloudbees/jenkins/support/actions/SupportChildAction/sidepanel.jelly
@@ -2,13 +2,14 @@
 <j:jelly xmlns:j="jelly:core" xmlns:l="/lib/layout">
   <l:side-panel>
     <l:tasks>
+      <j:set var="parent" value="${it.supportAction}"/>
 
       <l:task contextMenu="false"
               href="${rootURL}"
               icon="up"
               title="${%Back to Dashboard}"/>
 
-      <j:forEach var="action" items="${it.actions}">
+      <j:forEach var="action" items="${parent.actions}">
         <l:task icon="${h.getIconFilePath(action)}"
                 href="${rootURL}/${action.contextUrl}"
                 title="${action.displayName}"
