+++ src/main/resources/com/cloudbees/jenkins/support/actions/SupportGenerateBundleAction/index.properties
@@ -3,4 +3,3 @@ detail=It is best to include all of the following information in the support bun
   generating the bundle or edit the generated bundle to remove the specific information that you must exclude. \
   Each bundle is a simple ZIP file containing mostly plain text files and you can examine and/or modify the bundle \
   prior to sharing the bundle if you have any concerns about the information contained within.
-permissionPreReqs=Requires: {0}
