+++ src/main/resources/com/cloudbees/jenkins/support/actions/SupportGeneratedContentAction/index.properties
@@ -1,0 +1,1 @@
+ExistingSupportBundles=Existing Support bundles
